Moved containers in their own subdirectory
